pretty-print unicode unambiguously
As things are currently, if there's you see `0xff` (say) in the pretty-printed
grammar, it's ambiguous whether it means the byte 0xff or the rune 0xff
(represented by 0xc3 0xbf).

We change the output so that bytes and runes are unambiguously
represented using the same conventions used by Go's string quoting.

In doing so, it becomes clear that there's no real reason to distinguish
between bytes and UTF-8 input, because it's straightforward to
represent both internally. So we store input lexemes as the runeOrByte
type, and convert them appropriately on output, which removes
the need for the separate ParseUTF8 and ParseBinary functions
and the ErrMalformedUTF8 error.

Note: If we made it possible to have an empty grammar, then we
could parse all possible inputs and there would be no need
for an error return from Parse.